fix: search songs switching in player
